refactor(batch-exports): move the ClickHouse-reading producer into test support
`spmc.Producer` has no production callers. Every destination builds its record
batches with `pipeline.producer.Producer`, which reads from the internal S3
staging area. This one queries ClickHouse directly, and its only remaining
consumers are the destination test suites, which use it to build the rows they
expect and compare those against what the real pipeline produced.

Move it to `tests/temporal/utils/clickhouse_test_producer.py` and rename the
class to `ClickHouseTestProducer` so its role is clear at the call site. The
class body is otherwise unchanged. `default_fields` moves with it, since the
producer was its only caller — the live equivalent is
`batch_exports.default_fields`.

The `SELECT_FROM_*` templates deliberately stay in `temporal/sql/`. Each pairs
with an `EXPORT_TO_S3_FROM_*` template that production uses, three of those
pairs have already drifted, and splitting a pair across the test/production
boundary would make future drift harder to spot. The HTTP destination also
still reads them via `iter_records`.

Two knock-on changes:

- `internal_stage.py` imported `RecordBatchModel` from `spmc`, which was really
  a re-export of `record_batch_model`. Removing the producer left that import
  unused in `spmc`, so it now comes from the real home.
- The producer's own test asserted HogQL parameter substitution, which every
  destination suite already exercises via the `hogql_val` entries in its
  `TEST_MODELS`. Dropped rather than moved.

spmc.py drops from 791 to 500 lines and no longer imports any SQL.

Generated-By: PostHog Code
Task-Id: e2c0e744-189b-4ac0-a1f4-c999b0a0e071
