+++ products/batch_exports/backend/tests/temporal/destinations/bigquery/utils.py
@@ -18,7 +18,8 @@ from products.batch_exports.backend.temporal.destinations.bigquery_batch_export 
     get_our_google_cloud_credentials,
 )
 from products.batch_exports.backend.temporal.record_batch_model import SessionsRecordBatchModel
-from products.batch_exports.backend.temporal.spmc import Producer, RecordBatchQueue
+from products.batch_exports.backend.temporal.spmc import RecordBatchQueue
+from products.batch_exports.backend.tests.temporal.utils.clickhouse_test_producer import ClickHouseTestProducer
 from products.batch_exports.backend.tests.temporal.utils.records import get_record_batch_from_queue
 
 SKIP_IF_MISSING_GOOGLE_APPLICATION_CREDENTIALS = pytest.mark.skipif(
@@ -149,9 +150,9 @@ async def assert_clickhouse_records_in_bigquery(
     expected_records = []
     queue = RecordBatchQueue()
     if model_name == "sessions":
-        producer = Producer(model=SessionsRecordBatchModel(team_id))
+        producer = ClickHouseTestProducer(model=SessionsRecordBatchModel(team_id))
     else:
-        producer = Producer()
+        producer = ClickHouseTestProducer()
 
     for data_interval_start, data_interval_end in date_ranges:
         producer_task = await producer.start(
